Verification Class Updated!
In this dear commit, I added UserGetStatus and UserPostStatus, which it return details abouts the request such if had an error, type of method, if it was a success request, and so on

+++ src/main/java/com/blazeland/simplelogin/services/implementation/UserServiceImpl.java
@@ -3,6 +3,9 @@ package com.blazeland.simplelogin.services.implementation;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import com.blazeland.simplelogin.models.User;
+import com.blazeland.simplelogin.models.verification.UserGetStatus;
+import com.blazeland.simplelogin.models.verification.UserPostStatus;
+import com.blazeland.simplelogin.models.verification.UserVerification;
 import com.blazeland.simplelogin.repository.UserRepository;
 import com.blazeland.simplelogin.services.UserService;
 
@@ -12,37 +15,61 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class UserServiceImpl implements UserService {
     UserRepository userRepository;
+    UserPostStatus userPostStatus;
+    UserGetStatus userGetStatus;
 
     @Override
-    public boolean createUser(User user) {
-        // TODO miss many login
-        if (new UserVerification().verifyUser(user)) {
+    public UserPostStatus createUser(User user) {
+        userPostStatus = new UserVerification().verifyUser(user, userRepository);
+        userPostStatus.setMethod("POST");
+        if (userPostStatus.isSuccess()) {
             userRepository.save(user);
-            return true;
         }
-
-        return false;
+        return userPostStatus;
     }
 
     @Override
-    public boolean updateUser(User user) {
-        if (new UserVerification().verifyUser(user)) {
+    public UserPostStatus updateUser(User user) {
+        userPostStatus = new UserVerification().verifyUser(user, userRepository);
+        userPostStatus.setMethod("PUT");
+        if (userPostStatus.isSuccess()) {
             userRepository.save(user);
-            return true;
         }
-        return false;
+        return userPostStatus;
     }
 
     @Override
-    public boolean deleteUser(String userEmail, String userPassword) {
-        var userId = userRepository.findByUserEmailAndUserPassword(userEmail, userPassword).getUserId();
-        userRepository.deleteById(userId);
-        return true;
+    public UserGetStatus deleteUser(String userEmail, String userPassword) {
+        var userId = userRepository.findUserByEmailAndPassword(userEmail, userPassword).getUserId();
+        userGetStatus.setMethod("DELETE");
+        userGetStatus.setUser(null);
+        if (userId.equals(null)) {
+            userGetStatus.setSuccess(false);
+            userGetStatus.setDeleted(true);
+
+        } else {
+            userRepository.deleteById(userId);
+            userGetStatus.setSuccess(true);
+            userGetStatus.setDeleted(false);
+        }
+        return userGetStatus;
     }
 
     @Override
-    public User getUser(String userEmail, String userPassword) {
-        return userRepository.findByUserEmailAndUserPassword(userEmail, userPassword);
+    public UserGetStatus getUser(String userEmail, String userPassword) {
+        var user = userRepository.findUserByEmailAndPassword(userEmail, userPassword);
+        userGetStatus.setMethod("GET");
+        userGetStatus.setUser(null);
+        if (user.equals(null)) {
+            userGetStatus.setSuccess(false);
+            userGetStatus.setGet(false);
+        } else {
+            userGetStatus.setUser(user);
+            userGetStatus.setSuccess(true);
+            userGetStatus.setGet(true);
+        }
+        return userGetStatus;
+
     }
 
     @Override
@@ -50,4 +77,5 @@ public class UserServiceImpl implements UserService {
         return userRepository.findAll();
 
     }
+
 }
